+++ cli/src/bin/annis.rs
@@ -24,6 +24,13 @@ use std::{collections::BTreeSet, time::Duration};
 
 use anyhow::Result;
 
+#[cfg(not(target_env = "msvc"))]
+use tikv_jemallocator::Jemalloc;
+
+#[cfg(not(target_env = "msvc"))]
+#[global_allocator]
+static GLOBAL: Jemalloc = Jemalloc;
+
 #[derive(Helper, Hinter, Highlighter, Validator)]
 struct ConsoleHelper {
     known_commands: BTreeSet<String>,
@@ -360,14 +367,8 @@ impl AnnisRunner {
         for c in corpora {
             let desc = match c.load_status {
                 LoadStatus::NotLoaded => String::from("not loaded"),
-                LoadStatus::PartiallyLoaded(size) => format!(
-                    "partially loaded, {:.2} MB",
-                    size as f64 / (1024 * 1024) as f64
-                ),
-                LoadStatus::FullyLoaded(size) => format!(
-                    "fully loaded, {:.2} MB ",
-                    size as f64 / (1024 * 1024) as f64
-                ),
+                LoadStatus::PartiallyLoaded => String::from("partially loaded"),
+                LoadStatus::FullyLoaded => String::from("fully loaded"),
             };
             println!("{} ({})", c.name, desc);
         }